Extract arrayOf checker
